+++ tests/testthat/test-durations.R
@@ -10,6 +10,24 @@ test_that("new_difftime works as expected",{
   expect_that(diff, equals(y))
 })
 
+test_that("parsing works as expected", {
+  expect_equal(duration("1min 2sec 2secs 1H 2M 1d"),
+               duration(seconds = 4, minutes = 3, hours = 1, days = 1))
+  expect_equal(duration("day day"),
+               duration(days = 2))
+  expect_equal(duration("S M H d w"),
+               duration(seconds = 1, minutes = 1, hours = 1, days = 1, weeks = 1))
+})
+
+test_that("character comparison with durrations works as expected", {
+  expect_true("day" == duration(days = 1))
+  expect_true("2 days, 2 secs" == duration(days = 2, seconds = 2))
+  expect_true(duration("day 1s") >  duration(days = 1))
+  expect_true("day 1s" >  duration(days = 1))
+  expect_true("day 1S 2H" == duration(days = 1, seconds = 1, hours = 2))
+  expect_false("day 1S 2H" < duration(days = 1, hours = 2))
+})
+
 test_that("new_difftime handles vectors",{
   x <- as.POSIXct(c("2008-08-03 13:01:59", "2008-08-03 13:01:59"), tz = "UTC")
   y <- difftime(x + c(5 + 30*60 + 60*60 + 14*24*60*60,